lange file opgesplitst in componenten
